demo/requirements.txt
gradio==4.44.1
fastapi==0.115.0
starlette<0.41
huggingface_hub>=0.24,<1.0
plotly>=5.18
pandas>=2.0
PyYAML>=6.0
